backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0
aiosqlite==0.19.0
asyncmy==0.2.9

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis
redis==5.0.1

# WebSocket
websockets==12.0

# Security
bleach==6.1.0

# File handling
pillow==10.1.0
aiofiles==23.2.1
python-magic==0.4.27

# Utils
python-dotenv==1.0.0
python-slugify==8.0.1

